% Auto-generated: do not edit by hand
\name{'ydc'SampleComponent}

\alias{'ydc'SampleComponent}

\title{SampleComponent component}

\description{
Component description
}

\usage{
'ydc'SampleComponent(id=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

