% Auto-generated: do not edit by hand
\name{'ydc'TrackingLastComponent}

\alias{'ydc'TrackingLastComponent}

\title{TrackingLastComponent component}

\description{
Component description
}

\usage{
'ydc'TrackingLastComponent(children=NULL, id=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

